Get rid of duplicated logic
